pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "litemind"  # <-- Add the project name here
dynamic = ["version"]
description = "A wrapper API around LLM APIs and agentic AI framework"
authors = [{ name = "Loic A. Royer" }]
license = "BSD-3-Clause"
readme = "README.md"
requires-python = ">=3.8"
keywords = ["llm", "ai", "agents", "wrapper"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: BSD License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.6",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
]
dependencies = [
    "openai",
    "ollama",
    "anthropic",
    "google-generativeai",
    "pillow",
    "arbol",
    "ansicolors",
    "json-repair",
    "tiktoken",
    "pydantic",
    "pandas",
    "numpy",
    "chardet",
    "pyyaml",
    "absl-py",
    "scipy",
    "numpy",
]

[tool.hatch.build.targets.sdist]
ignore-vcs = true
packages = ["src/litemind"]
include = [
    "**",
    "README.md",
    "LICENSE",
]
exclude = [
    "**/tests/**",
    "**/.DS_Store",
    "*.egg-info",
    "**/sandbox/**",
    "src/litemind/media/**"
]

[tool.hatch.version]
path = "src/litemind/__init__.py"

[project.urls]
Homepage = "https://github.com/royerlab/litemind"
Issues = "https://github.com/royerlab/litemind/issues"

# Development dependencies:
[project.optional-dependencies]
dev = [
    "hatch",
    "pytest",
    "pytest-html",
    "pytest-cov",
    "pytest-md-report",
    "black",
    "flake8",
    "mypy",
    "isort",
]

# Optional dependencies:
rag = [
    "fastembed",
    "chromadb",
    "qdrant-client"
]
whisper = [
    "openai-whisper"
]
documents = [
    "pymupdf",
    "pymupdf4llm",
    "openpyxl"
]
tables = [
    "tabulate"
]
videos = [
    "ffmpeg-python",
    "imageio[ffmpeg]",
]
audio = [
    "soundfile"
]


# Commands:
[project.scripts]
litemind = "litemind.tools.litemind_tools:main"

[tool.hatch.envs.hatch-test]
default-args = ["src/"]
exclude = [
    "sandbox",
]
extra-dependencies = [
    "qdrant-client",
    "chromadb",
    "openai-whisper",
    "pymupdf",
    "pymupdf4llm",
    "openpyxl",
    "pandas",
    "numpy",
    "chardet",
    "tabulate",
    "fastembed",
    "ffmpeg-python",
    "imageio[ffmpeg]",
    "pytest-html",
    "pytest-cov",
    "pytest-md-report",
    "pypdf",
    "soundfile"
]


